// src/lib/types.ts
export interface PlansI {
  name: string;
  for: string;
  capital: number;
  advantage: number;
  fees: string;
  leverageCapacity: string;
  lotSize: string;
  limit: number;
  openPositionCapacity: string;
  threshold: number;
  activation: number;
  policy: number;
  risk: string;
  options: string;
}